src/third_party/blink/web_tests/accessibility/aria-hidden.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/run-after-layout-and-paint.js"></script>
</head>
<body id="body">
<h1 aria-hidden="true">h1 <b>test</b></h1>
<h2 aria-hidden="false">h2</h2>
<p id="description"></p>
<div id="console"></div>

<script>
test_after_layout_and_paint((t) => {
    var body = document.getElementById("body");
    body.focus();
    var h2 = accessibilityController.focusedElement.childAtIndex(0);
    assert_equals(h2.name, "h2");
}, "This tests that the aria-hidden attribute works correctly with accessibility. The H1 element (and its children) should not appear in the AX hierarchy. The H2 element should be the first child");


</script>

</body>
</html>
